fix: disable run region when only select open code
