+++ client/src/commands/run.ts
@@ -2,9 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 import {
   EventEmitter,
-  Position,
   ProgressLocation,
-  Selection,
   Uri,
   commands,
   l10n,
@@ -20,14 +18,8 @@ import { isOutputHtmlEnabled } from "../components/Helper/SettingHelper";
 import { LogFn as LogChannelFn } from "../components/LogChannel";
 import { showResult } from "../components/ResultPanel";
 import { OnLogFn, RunResult, getSession } from "../connection";
+import { getSelectedRegions } from "../utils/utils";
 import { profileConfig, switchProfile } from "./profile";
-
-interface FoldingBlock {
-  startLine: number;
-  startCol: number;
-  endLine: number;
-  endCol: number;
-}
 
 let running = false;
 
@@ -69,57 +61,6 @@ function getCode(selected = false, uri?: Uri): string {
     code = assign_SASProgramFile(code, codeFile);
   }
   return wrapCodeWithOutputHtml(code);
-}
-
-async function getSelectedRegions(
-  client: BaseLanguageClient,
-): Promise<Selection[]> {
-  const result: string[] = [];
-
-  async function pushBlock(line: number, col: number) {
-    const block = await client.sendRequest<FoldingBlock>(
-      "sas/getFoldingBlock",
-      {
-        textDocument: { uri: window.activeTextEditor.document.uri.toString() },
-        line,
-        col,
-      },
-    );
-    if (block) {
-      const start = doc.offsetAt(new Position(block.startLine, block.startCol));
-      const end = doc.offsetAt(new Position(block.endLine, block.endCol));
-      const key = `${start}-${end}`;
-      if (result.indexOf(key) === -1) {
-        result.push(key);
-      }
-      return end;
-    }
-  }
-
-  const editor = window.activeTextEditor;
-  const doc = editor.document;
-  for (const selection of editor.selections) {
-    const start = doc.offsetAt(selection.start);
-    let end = doc.offsetAt(selection.end);
-    const selectedText = doc.getText(selection);
-    if (selectedText.endsWith("\n")) {
-      --end;
-    }
-    for (let i = start; i <= end; i++) {
-      const pos = doc.positionAt(i);
-      const blockEnd = await pushBlock(pos.line, pos.character);
-      if (blockEnd && blockEnd > i) {
-        i = blockEnd;
-      }
-    }
-  }
-  return result.map((key) => {
-    const [start, end] = key.split("-");
-    return new Selection(
-      doc.positionAt(parseInt(start)),
-      doc.positionAt(parseInt(end)),
-    );
-  });
 }
 
 async function runCode(selected?: boolean, uri?: Uri) {
